Apply `title` from url on page load
