+++ .changeset/green-plants-cheer.md
@@ -1,0 +1,8 @@
+---
+'playroom': patch
+---
+
+Apply `title` from url on page load
+
+Previously the document `title` would only update when the frames panel is open.
+The title is now correctly reflected from the url on page load.
