+++ src/Playroom/FramesPanel/FramesPanel.tsx
@@ -5,25 +5,10 @@ import { ToolbarPanel } from '../ToolbarPanel/ToolbarPanel';
 import { StoreContext } from '../../StoreContext/StoreContext';
 import { Stack } from '../Stack/Stack';
 import { Text } from '../Text/Text';
-import { Helmet } from 'react-helmet';
 import { Inline } from '../Inline/Inline';
 import { Box } from '../Box/Box';
 
 import * as styles from './FramesPanel.css';
-
-const getTitle = (title: string | undefined) => {
-  if (title) {
-    return `${title} | Playroom`;
-  }
-
-  const configTitle = window?.__playroomConfig__.title;
-
-  if (configTitle) {
-    return `${configTitle} | Playroom`;
-  }
-
-  return 'Playroom';
-};
 
 interface FramesPanelProps {
   availableWidths: number[];
@@ -104,93 +89,84 @@ export default ({ availableWidths, availableThemes }: FramesPanelProps) => {
   const hasFilteredThemes =
     visibleThemes.length > 0 && visibleThemes.length <= availableThemes.length;
 
-  const displayedTitle = getTitle(title);
-
   return (
-    <>
-      {title === undefined ? null : (
-        <Helmet>
-          <title>{displayedTitle}</title>
-        </Helmet>
-      )}
-      <ToolbarPanel>
-        <Stack space="xxxlarge">
-          <label>
-            <Stack space="small">
-              <Heading level="3">Title</Heading>
-              <input
-                type="text"
-                id="playroomTitleField"
-                placeholder="Enter a title for this Playroom..."
-                className={styles.textField}
-                value={title}
-                onChange={(e) =>
-                  dispatch({
-                    type: 'updateTitle',
-                    payload: { title: e.target.value },
-                  })
-                }
-              />
-            </Stack>
-          </label>
+    <ToolbarPanel>
+      <Stack space="xxxlarge">
+        <label>
+          <Stack space="small">
+            <Heading level="3">Title</Heading>
+            <input
+              type="text"
+              id="playroomTitleField"
+              placeholder="Enter a title for this Playroom..."
+              className={styles.textField}
+              value={title}
+              onChange={(e) =>
+                dispatch({
+                  type: 'updateTitle',
+                  payload: { title: e.target.value },
+                })
+              }
+            />
+          </Stack>
+        </label>
 
+        <Stack space="xlarge">
+          <FrameHeading
+            showReset={hasFilteredWidths}
+            onReset={() => dispatch({ type: 'resetVisibleWidths' })}
+          >
+            Widths
+          </FrameHeading>
+          {availableWidths.map((option) => (
+            <FrameOption
+              key={option}
+              option={option}
+              selected={hasFilteredWidths && visibleWidths.includes(option)}
+              visible={visibleWidths}
+              onChange={(newWidths) => {
+                if (newWidths) {
+                  dispatch({
+                    type: 'updateVisibleWidths',
+                    payload: { widths: newWidths },
+                  });
+                } else {
+                  dispatch({ type: 'resetVisibleWidths' });
+                }
+              }}
+            />
+          ))}
+        </Stack>
+
+        {hasThemes ? (
           <Stack space="xlarge">
             <FrameHeading
-              showReset={hasFilteredWidths}
-              onReset={() => dispatch({ type: 'resetVisibleWidths' })}
+              showReset={hasFilteredThemes}
+              onReset={() => dispatch({ type: 'resetVisibleThemes' })}
             >
-              Widths
+              Themes
             </FrameHeading>
-            {availableWidths.map((option) => (
+            {availableThemes.map((option) => (
               <FrameOption
                 key={option}
                 option={option}
-                selected={hasFilteredWidths && visibleWidths.includes(option)}
-                visible={visibleWidths}
-                onChange={(newWidths) => {
-                  if (newWidths) {
+                selected={hasFilteredThemes && visibleThemes.includes(option)}
+                visible={visibleThemes}
+                onChange={(newThemes) => {
+                  if (newThemes) {
                     dispatch({
-                      type: 'updateVisibleWidths',
-                      payload: { widths: newWidths },
+                      type: 'updateVisibleThemes',
+                      payload: { themes: newThemes },
                     });
                   } else {
-                    dispatch({ type: 'resetVisibleWidths' });
+                    dispatch({ type: 'resetVisibleThemes' });
                   }
                 }}
               />
             ))}
           </Stack>
-
-          {hasThemes ? (
-            <Stack space="xlarge">
-              <FrameHeading
-                showReset={hasFilteredThemes}
-                onReset={() => dispatch({ type: 'resetVisibleThemes' })}
-              >
-                Themes
-              </FrameHeading>
-              {availableThemes.map((option) => (
-                <FrameOption
-                  key={option}
-                  option={option}
-                  selected={hasFilteredThemes && visibleThemes.includes(option)}
-                  visible={visibleThemes}
-                  onChange={(newThemes) => {
-                    if (newThemes) {
-                      dispatch({
-                        type: 'updateVisibleThemes',
-                        payload: { themes: newThemes },
-                      });
-                    } else {
-                      dispatch({ type: 'resetVisibleThemes' });
-                    }
-                  }}
-                />
-              ))}
-            </Stack>
-          ) : null}
-        </Stack>
-      </ToolbarPanel>
-    </>
+        ) : null}
+      </Stack>
+    </ToolbarPanel>
   );
 };
